+++ ruoyi-modules/ruoyi-system/src/main/java/com/ruoyi/system/service/ISysPostService.java
@@ -1,6 +1,7 @@
 package com.ruoyi.system.service;
 
 import java.util.List;
+
 import com.ruoyi.system.domain.SysPost;
 
 /**
@@ -39,7 +40,7 @@ public interface ISysPostService
      * @param userId 用户ID
      * @return 选中岗位ID列表
      */
-    public List<Long> selectPostListByUserId(Long userId);
+    public List<Integer> selectPostListByUserId(Long userId);
 
     /**
      * 校验岗位名称
@@ -78,6 +79,7 @@ public interface ISysPostService
      * 
      * @param postIds 需要删除的岗位ID
      * @return 结果
+     * @throws Exception 异常
      */
     public int deletePostByIds(Long[] postIds);
